// src/types/typography.d.ts
export type FontWeight = 'light' | 'regular' | 'medium' | 'semibold' | 'bold';
export type LineHeight = 'tight' | 'snug' | 'normal' | 'relaxed' | 'loose';
export type LetterSpacing = 'tight' | 'normal' | 'wide' | 'wider';
export type TextTransform = 'uppercase' | 'lowercase' | 'capitalize' | 'normal';
export type TextAlign = 'left' | 'center' | 'right' | 'justify';